spc plus LotID 세팅 Global - Lotid PLC 데이터

+++ NotchingGradeInsp/CImageProcess.cpp
@@ -9929,7 +9929,7 @@ BOOL CImageProcess::GetBrightAvg_PetCheck(BYTE* pOrgImg, int nImageWidth, int nI
 
 
 
-	int nSize = vstPetInfo->size();
+	int nSize = (int)vstPetInfo->size();
